requirements.txt
streamlit==1.44.1
pillow==11.1.0
agno>=2.2.10
google-genai>=1.9.0
duckduckgo-search
python-dotenv